widgets for loader were split into separate files into _widgets folder
