+++ avalon/tools/_widgets/__init__.py
@@ -1,0 +1,11 @@
+from ...vendor.Qt import QtCore, QtWidgets, QtGui
+from ...vendor import qtawesome
+from ... import api, pipeline, io, style
+
+
+from .widget_silo_tab import SiloTabWidget
+from .widget_asset import AssetWidget
+from .text_edit_version import VersionTextEdit
+from .widget_family_list import FamilyListWidget
+from .widget_subset import SubsetWidget
+from .widget_version import VersionWidget
