+++ avalon/tools/_widgets/widget_silo_tab.py
@@ -1,0 +1,143 @@
+from . import QtCore, QtWidgets, QtGui
+
+
+class SiloTabWidget(QtWidgets.QTabBar):
+    """Silo widget
+
+    Allows to add a silo, with "+" tab.
+
+    Note:
+        When no silos are present an empty stub silo is added to
+        use as the "blank" tab to start on, so the + tab becomes
+        clickable.
+
+    """
+
+    silo_changed = QtCore.Signal(str)
+    silo_added = QtCore.Signal(str)
+
+    def __init__(self, parent=None):
+        super(SiloTabWidget, self).__init__(parent=parent)
+        self._previous_tab_index = -1
+        self.set_silos([])
+
+        self.setContentsMargins(0, 0, 0, 0)
+        self.setFixedHeight(28)
+        font = QtGui.QFont()
+        font.setBold(True)
+        self.setFont(font)
+
+        self.currentChanged.connect(self.on_tab_changed)
+
+    def on_tab_changed(self, index):
+
+        if index == self._previous_tab_index:
+            return
+
+        # If it's the last tab
+        num = self.count()
+        if index == num - 1:
+            self.on_add_silo()
+            self.setCurrentIndex(self._previous_tab_index)
+            return
+
+        silo = self.tabText(index)
+        self.silo_changed.emit(silo)
+
+        # Store for the next calls
+        self._previous_tab_index = index
+
+    def clear(self):
+        """Removes all tabs.
+
+        Implemented similar to `QTabWidget.clear()`
+
+        """
+        for i in range(self.count()):
+            self.removeTab(0)
+
+    def set_silos(self, silos):
+
+        current_silo = self.get_current_silo()
+
+        if not silos:
+            # Add an emtpy stub tab to start on.
+            silos = [""]
+
+        # Populate the silos without emitting signals
+        self.blockSignals(True)
+        self.clear()
+        for silo in sorted(silos):
+            self.addTab(silo)
+
+        # Add the "+" tab
+        self.addTab("+")
+
+        self.set_current_silo(current_silo)
+        self.blockSignals(False)
+
+        # Assume the current index is "fine"
+        self._previous_tab_index = self.currentIndex()
+
+        # Only emit a silo changed signal if the new signal
+        # after refresh is not the same as prior to it (e.g.
+        # when the silo was removed, or alike.)
+        if current_silo != self.get_current_silo():
+            self.currentChanged.emit(self.currentIndex())
+
+    def set_current_silo(self, silo):
+        """Set the active silo by name or index.
+
+        Args:
+            silo (str or int): The silo name or index.
+            emit (bool): Whether to emit the change signals
+
+        """
+
+        # Already set
+        if silo == self.get_current_silo():
+            return
+
+        # Otherwise change the silo box to the name
+        for i in range(self.count()):
+            text = self.tabText(i)
+            if text == silo:
+                self.setCurrentIndex(i)
+                break
+
+    def get_current_silo(self):
+        index = self.currentIndex()
+        return self.tabText(index)
+
+    def on_add_silo(self):
+        silo, state = QtWidgets.QInputDialog.getText(
+            self, "Silo name", "Create new silo:"
+        )
+        if not state or not silo:
+            return
+
+        self.add_silo(silo)
+
+    def get_silos(self):
+        """Return the currently available silos"""
+
+        # Ignore first tab if empty
+        # Ignore the last tab because it is the "+" tab
+        silos = []
+        for i in range(self.count() - 1):
+            label = self.tabText(i)
+            if i == 0 and not label:
+                continue
+            silos.append(label)
+        return silos
+
+    def add_silo(self, silo):
+
+        # Add the silo to tab
+        silos = self.get_silos()
+        silos.append(silo)
+        silos = list(set(silos))  # ensure unique
+        self.set_silos(silos)
+        self.silo_added.emit(silo)
+
+        self.set_current_silo(silo)
